+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-# v.0.1.10
+# v0.1.11
+
+- Services timeout is set to 3000 ms and moved to config
+
+Bug Fixed:
+- Checking required fields if set(#1)
+
+
+# v0.1.10
 
 - `req.setTimeout` is set to `ip-api` method
 - Debug Project name is set to GeoSearch; for using debug mode: `DEBUG=GeoSearch:*`